+++ battleservice/src/services/battle/scene/consts/MsgPosScaleRate.go
@@ -1,5 +1,5 @@
 package consts
 
 const (
-	MsgPosScaleRate float64 = 100 // 服务器内部坐标 * MsgPosScaleRate = 发送给客户端的坐标
+	MsgPosScaleRate float32 = 100 // 服务器内部坐标 * MsgPosScaleRate = 发送给客户端的坐标
 )
